add detail/macro_{un}scope.hpp
