+++ Makefile
@@ -2,7 +2,9 @@
 
 SRCDIR = src
 SRCS = ${SRCDIR}/json.hpp \
-			 ${SRCDIR}/json_fwd.hpp
+			 ${SRCDIR}/json_fwd.hpp \
+			 ${SRCDIR}/detail/macro_scope.hpp \
+			 ${SRCDIR}/detail/macro_unscope.hpp
 
 # main target
 all:
